# API/swagger.yaml
openapi: 3.0.0
info:
  title: Fruity Tunes API
  version: 1.0.0
  description: API per gestionar cançons, jugadors i puntuacions del joc de ritme Fruity Tunes
components:
  schemas:
    ChartNote:
      type: object
      properties:
        inputBeat:
          type: number
          example: 1.5
          description: Beat on apareix la nota
        inputKey:
          type: string
          enum:
            - left
            - right
          example: left
    Song:
      type: object
      properties:
        songId:
          type: uuid
          example: 019ebbb2-31cb-7879-a4a6-9aa3e6b89ff3
        songTitle:
          type: string
          example: Must, Be Nice
        bpm:
          type: number
          example: 140
        audioFile:
          type: string
          example: grapes1_must-be-nice.mp3
        offset:
          type: number
          example: 0.05
          description: Offset en segons
        chart:
          type: array
          items:
            $ref: '#/components/schemas/ChartNote'
    Score:
      type: object
      properties:
        scoreId:
          type: uuid
          example: 019ebbb2-31cb-7250-925d-c5f8341fc4c2
        highscore:
          type: integer
          example: 98500
        maxCombo:
          type: integer
          example: 312
        rank:
          type: string
          example: S
    LevelEntry:
      type: object
      properties:
        songId:
          type: uuid
          example: 019ebbb2-31cb-7879-a4a6-9aa3e6b89ff3
        scores:
          type: array
          maxItems: 5
          description: Fins a 5 millors puntuacions, en ordre descendent
          items:
            $ref: '#/components/schemas/Score'
    Player:
      type: object
      properties:
        playerId:
          type: uuid
          example: 019ebbb2-31cb-7a60-8e1d-ce9956756667
        playerName:
          type: string
          example: RhythmMaster
        levelScores:
          type: array
          items:
            $ref: '#/components/schemas/LevelEntry'
    LeaderboardEntry:
      type: object
      description: Fila de la taula de classificacions
      properties:
        playerId:
          type: uuid
          example: 019ebbb2-31cb-7a60-8e1d-ce9956756667
        playerName:
          type: string
          example: RhythmMaster
        highscore:
          type: integer
          example: 98500
        maxCombo:
          type: integer
          example: 312
        rank:
          type: string
          example: S
    Error:
      type: object
      properties:
        error:
          type: string
          example: Cançó no trobada
paths:
  /songs:
    get:
      summary: Llista totes les cançons
      tags:
        - Songs
      responses:
        '200':
          description: Array amb totes les cançons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Song'
    post:
      summary: Afegeix una nova cançó
      tags:
        - Songs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - songId
                - songTitle
                - bpm
                - audioFile
              properties:
                songId:
                  type: uuid
                  example: 019ebbb2-31cb-7879-a4a6-9aa3e6b89ff3
                songTitle:
                  type: string
                  example: Must
                  Be Nice: null
                bpm:
                  type: number
                  example: 120
                audioFile:
                  type: string
                  example: grapes1_must-be-nice.mp3
                offset:
                  type: number
                  example: 0
                chart:
                  type: array
                  items:
                    $ref: '#/components/schemas/ChartNote'
      responses:
        '201':
          description: Cançó creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Song'
        '400':
          description: Falten camps obligatoris
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Ja existeix una cançó amb aquest songId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /songs/{songId}:
    get:
      summary: Obté una cançó per ID
      tags:
        - Songs
      parameters:
        - in: path
          name: songId
          required: true
          schema:
            type: uuid
          example: 019ebbb2-31cb-7879-a4a6-9aa3e6b89ff3
      responses:
        '200':
          description: Dades de la cançó
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Song'
        '404':
          description: Cançó no trobada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /players:
    get:
      summary: Llista tots els jugadors
      description: Retorna una llista de tots els jugadors registrats al sistema, sense incloure les seves puntuacions
      tags:
        - Players
      responses:
        '200':
          description: Llista de jugadors obtinguda correctament
          content:
            application/json:
              schema:
                type: object
                properties:
                  players:
                    type: array
                    items:
                      type: object
                      properties:
                        playerId:
                          type: string
                          format: uuid
                          example: 019ebbb2-31cb-7a60-8e1d-ce9956756667
                          description: ID únic del jugador
                        playerName:
                          type: string
                          example: RhythmMaster
                          description: Nom del jugador
        '500':
          description: Error intern del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Error en obtenir els jugadors
    post:
      summary: Registra un nou jugador
      tags:
        - Players
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - playerName
              properties:
                playerName:
                  type: string
                  example: ComboKing
      responses:
        '201':
          description: Jugador creat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        '400':
          description: Falten camps obligatoris
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Ja existeix un jugador amb aquest playerId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /players/{playerId}:
    get:
      summary: Obté un jugador amb totes les seves puntuacions
      tags:
        - Players
      parameters:
        - in: path
          name: playerId
          required: true
          schema:
            type: uuid
          example: 019ebbb2-31cb-7a60-8e1d-ce9956756667
      responses:
        '200':
          description: Dades del jugador i puntuacions agrupades per cançó (fins a 5 puntuacions per cançó)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        '404':
          description: Jugador no trobat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /scores/{playerId}:
    post:
      summary: Envia una nova puntuació d'un jugador en una cançó
      description: |
        Afegeix la puntuació a l'historial d'aquesta cançó pel jugador. Es conserven únicament les **5 millors puntuacions** ordenades per highscore descendent Si el jugador ja té 5 puntuacions i la nova no sepra la pitjor, es descarta. Si el jugador ya tiene 5 scores y el nuevo no supera al peor, se descarta.
      tags:
        - Players
      parameters:
        - in: path
          name: playerId
          required: true
          schema:
            type: uuid
          example: 019ebbb2-31cb-7a60-8e1d-ce9956756667
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - playerId
                - highscore
                - maxCombo
                - rank
              properties:
                playerId:
                  type: uuid
                  example: 019ebbb2-31cb-7a60-8e1d-ce9956756667
                highscore:
                  type: integer
                  example: 105000
                maxCombo:
                  type: integer
                  example: 400
                rank:
                  type: string
                  example: SS
      responses:
        '200':
          description: Jugador actualitzat amb el nou historial de puntuacions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        '400':
          description: Falten camps obligatoris
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Jugador o cançó no trobats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /scores/{songId}:
    get:
      summary: Puntuacions de tots els jugadors en una cançó
      description: Retorna la millor puntuació de cada jugador per aquesta cançó, ordenats de major a menor.
      tags:
        - Scores
      parameters:
        - in: path
          name: songId
          required: true
          schema:
            type: uuid
          example: 019ebbb2-31cb-7879-a4a6-9aa3e6b89ff3
      responses:
        '200':
          description: Llista de puntuacions ordenada
          content:
            application/json:
              schema:
                type: object
                properties:
                  songId:
                    type: uuid
                    example: 019ebbb2-31cb-7879-a4a6-9aa3e6b89ff3
                  scores:
                    type: array
                    items:
                      $ref: '#/components/schemas/LeaderboardEntry'
        '404':
          description: Cançó no trobada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
tags: []